Create summary report api
